Fix undefined variable issue when no page specified
When calling the index page without specifying a page (when you first arrive at the site for instance) the if statement goes to a page reload, however it doesn't exit, so the subsequent code still runs, trying to include page content. The page hasn't been specified though so it causes an undefined variable error to be logged each time.

While fixing the above issue I realised the code could be simplified to remove the page reload completely and just display the home page, which is all that appears to be happening anyway.